move-deprecated-classes
