+++ legend-depot-artifacts-repository-api/src/main/java/org/finos/legend/depot/artifacts/repository/RepositoryModule.java
@@ -57,7 +57,7 @@ public class RepositoryModule extends PrivateModule
         metricsHandler.registerGauge(MISSING_REPO_VERSIONS, MISSING_REPO_VERSIONS);
         metricsHandler.registerGauge(MISSING_STORE_VERSIONS, MISSING_STORE_VERSIONS);
         metricsHandler.registerGauge(REPO_EXCEPTIONS, REPO_EXCEPTIONS);
-        schedulesFactory.register(REPOSITORY_METRICS_SCHEDULE, LocalDateTime.now().plusMinutes(5), 300000, false,repositoryServices::findVersionsMismatches);
+        schedulesFactory.register(REPOSITORY_METRICS_SCHEDULE, LocalDateTime.now().plusMinutes(1), 300000, false,repositoryServices::findVersionsMismatches);
         return true;
     }
 }
