+++ legend-depot-server/src/main/java/org/finos/legend/depot/server/resources/ProjectsResource.java
@@ -15,16 +15,23 @@
 
 package org.finos.legend.depot.server.resources;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.lang.builder.EqualsBuilder;
+import org.apache.commons.lang.builder.HashCodeBuilder;
 import org.finos.legend.depot.domain.CoordinateData;
+import org.finos.legend.depot.domain.VersionedData;
+import org.finos.legend.depot.domain.project.ProjectVersion;
 import org.finos.legend.depot.domain.project.StoreProjectData;
-import org.finos.legend.depot.domain.project.ProjectData;
 import org.finos.legend.depot.domain.project.StoreProjectVersionData;
-import org.finos.legend.depot.domain.project.ProjectVersionProperty;
+import org.finos.legend.depot.domain.version.VersionValidator;
 import org.finos.legend.depot.services.api.projects.ProjectsService;
 import org.finos.legend.depot.tracing.resources.BaseResource;
 import org.finos.legend.depot.tracing.resources.ResourceLoggingAndTracing;
+import org.finos.legend.sdlc.domain.model.version.VersionId;
 
 import javax.inject.Inject;
 import javax.ws.rs.GET;
@@ -32,6 +39,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -75,10 +83,9 @@ public class ProjectsResource extends BaseResource
     @Path("/projects/versions/all/projectData")
     @ApiOperation(ResourceLoggingAndTracing.GET_ALL_PROJECTS)
     @Produces(MediaType.APPLICATION_JSON)
-    public List<ProjectDataDTO> getProjectsWithCoordinates()
+    public List<StoreProjectData> getProjectsWithCoordinates()
     {
-        return handle(ResourceLoggingAndTracing.GET_ALL_PROJECTS, () ->
-            projectApi.getAllProjectCoordinates().stream().map(pc -> new ProjectDataDTO(pc.getProjectId(), pc.getGroupId(), pc.getArtifactId())).collect(Collectors.toList()));
+        return handle(ResourceLoggingAndTracing.GET_ALL_PROJECTS, () -> projectApi.getAllProjectCoordinates());
     }
 
     @GET
@@ -117,24 +124,219 @@ public class ProjectsResource extends BaseResource
         return projectData;
     }
 
-    public class ProjectDataDTO extends CoordinateData
-    {
-        private String projectId;
 
-        public ProjectDataDTO()
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @Deprecated
+    public static final class ProjectData extends CoordinateData
+    {
+        @JsonProperty
+        private String projectId;
+        @JsonProperty
+        private List<String> versions = new ArrayList<>();
+        @JsonProperty
+        private List<ProjectVersionDependency> dependencies = new ArrayList<>();
+        @JsonProperty
+        private List<ProjectVersionProperty> properties = new ArrayList<>();
+
+        public ProjectData()
         {
-            super();
         }
 
-        public ProjectDataDTO(String projectId, String groupId, String artifactId)
+        public ProjectData(String projectId, String groupId, String artifactId)
         {
             super(groupId, artifactId);
             this.projectId = projectId;
         }
 
+        @JsonIgnore
+        public String getId()
+        {
+            return "";
+        }
+
         public String getProjectId()
         {
             return projectId;
+        }
+
+        public List<String> getVersions()
+        {
+            Collections.sort(versions);
+            return versions;
+        }
+
+        public void addVersion(String versionId)
+        {
+            if (!versionId.equals(VersionValidator.MASTER_SNAPSHOT) && !this.getVersions().contains(versionId))
+            {
+                this.versions.add(versionId);
+            }
+        }
+
+        @Override
+        public boolean equals(Object obj)
+        {
+            return EqualsBuilder.reflectionEquals(this, obj);
+        }
+
+        @Override
+        public int hashCode()
+        {
+            return HashCodeBuilder.reflectionHashCode(this);
+        }
+
+        @JsonProperty("latestVersion")
+        public String getLatestVersionAsString()
+        {
+            Optional<VersionId> latest = getLatestVersion();
+            return latest.map(VersionId::toVersionIdString).orElse(null);
+        }
+
+        @JsonIgnore
+        public Optional<VersionId> getLatestVersion()
+        {
+            if (versions != null && !versions.isEmpty())
+            {
+                List<VersionId> versionIds = versions.stream().map(VersionId::parseVersionId).collect(Collectors.toList());
+                return versionIds.stream().max(VersionId::compareTo);
+            }
+            return Optional.empty();
+        }
+
+        public List<ProjectVersionDependency> getDependencies()
+        {
+            return dependencies;
+        }
+
+        @JsonIgnore
+        public List<ProjectVersionDependency> getDependencies(String version)
+        {
+            return dependencies.stream().filter(dependency -> dependency.getVersionId().equals(version)).collect(Collectors.toList());
+        }
+
+        public List<ProjectVersionProperty> getPropertiesForProjectVersionID(String projectVersionId)
+        {
+            return properties.stream().filter(property -> property.getProjectVersionId().equals(projectVersionId)).collect(Collectors.toList());
+        }
+
+        public void addDependencies(List<ProjectVersionDependency> dependencies)
+        {
+            this.dependencies.addAll(dependencies);
+        }
+
+        public void setDependencies(List<ProjectVersionDependency> dependencies)
+        {
+            this.dependencies = dependencies;
+        }
+
+
+        public List<ProjectVersionProperty> getProperties()
+        {
+            return properties;
+        }
+
+        public void setProperties(List<ProjectVersionProperty> properties)
+        {
+            this.properties = properties;
+        }
+
+        public void addProperties(List<ProjectVersionProperty> propertyList)
+        {
+            propertyList.stream().filter(property -> !properties.contains(property)).forEach(property -> this.properties.add(property));
+        }
+
+        @JsonIgnore
+        public Optional<String> getVersion(String versionId)
+        {
+            return this.versions.stream().filter(v -> v.equals(versionId)).findFirst();
+        }
+
+        public static class ProjectVersionDependency extends VersionedData
+        {
+            private ProjectVersion dependency;
+
+            public ProjectVersionDependency()
+            {
+            }
+
+            public ProjectVersionDependency(String groupid, String artifactId, String versionId, ProjectVersion dep)
+            {
+                super(groupid, artifactId, versionId);
+                this.dependency = dep;
+            }
+
+            public ProjectVersion getDependency()
+            {
+                return dependency;
+            }
+
+            public void setDependency(ProjectVersion dependency)
+            {
+                this.dependency = dependency;
+            }
+
+            @Override
+            public boolean equals(Object obj)
+            {
+                return EqualsBuilder.reflectionEquals(this, obj);
+            }
+
+            @Override
+            public int hashCode()
+            {
+                return HashCodeBuilder.reflectionHashCode(this);
+            }
+        }
+    }
+
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    @Deprecated
+    public static final class ProjectVersionProperty
+    {
+        @JsonProperty
+        private String propertyName;
+        @JsonProperty
+        private String value;
+        @JsonProperty
+        private String projectVersionId;
+
+        public ProjectVersionProperty()
+        {
+
+        }
+
+        public ProjectVersionProperty(String propertyName, String value, String projectVersionId)
+        {
+            this.propertyName = propertyName;
+            this.value = value;
+            this.projectVersionId = projectVersionId;
+        }
+
+        public String getProjectVersionId()
+        {
+            return projectVersionId;
+        }
+
+        public String getPropertyName()
+        {
+            return propertyName;
+        }
+
+        public String getValue()
+        {
+            return value;
+        }
+
+        @Override
+        public boolean equals(Object obj)
+        {
+            return EqualsBuilder.reflectionEquals(this, obj);
+        }
+
+        @Override
+        public int hashCode()
+        {
+            return HashCodeBuilder.reflectionHashCode(this);
         }
     }
 }
